requirements.txt
python-dotenv
numpy 
pandas 
flask 
streamlit
certifi
pymongo[srv]==3.6
pymongo
scikit-learn
dill
pyaml

-e .